refactor: 공통 컴포넌트 스타일, 타입 파일 분리
