+++ src/components/Card/Card.types.ts
@@ -1,0 +1,6 @@
+import { ReactNode } from "react";
+
+export interface IUserCardProps {
+  children?: ReactNode;
+  onClick?: () => void;
+}
