+++ src/components/Feed/Feed.types.ts
@@ -1,0 +1,7 @@
+export interface FeedProps {
+  cover_image_url: string;
+  created_at: string;
+  id: string;
+  title: string;
+  user_id: string;
+}
