+++ src/components/Filler/Filler.types.ts
@@ -1,0 +1,4 @@
+export interface StyleProps {
+  width?: number;
+  height?: number;
+}
