+++ src/components/Modal/Modal.tsx
@@ -1,6 +1,12 @@
-import styled from "styled-components";
+import React from "react";
+import * as S from "./Modal.styles";
 
-const Modal = ({ onClose, children }: any) => {
+interface Props {
+  onClose: () => void;
+  children: React.ReactNode;
+}
+
+const Modal = ({ onClose, children }: Props) => {
   return (
     <S.ModalOverlay onClick={onClose}>
       <S.ModalContent onClick={(e) => e.stopPropagation()}>
@@ -8,29 +14,6 @@ const Modal = ({ onClose, children }: any) => {
       </S.ModalContent>
     </S.ModalOverlay>
   );
-};
-
-const S = {
-  ModalContent: styled.div`
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    background-color: #ffffff;
-    padding: 20px;
-    border-radius: 8px;
-    box-shadow: 0 4px 8px rgba(0, 0, 0, 0.1);
-    z-index: 3;
-  `,
-  ModalOverlay: styled.div`
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background-color: rgba(0, 0, 0, 0.3);
-    z-index: 4;
-  `,
 };
 
 export default Modal;
