+++ src/components/Profile/Profile.types.ts
@@ -1,0 +1,5 @@
+export interface StyleProps {
+  src: string;
+  width?: number;
+  height?: number;
+}
