+++ src/components/SearchBar/SearchBar.tsx
@@ -1,13 +1,8 @@
-import { styled } from "styled-components";
 import CommonInput from "../Input/Input";
-import { media } from "../../style/theme";
+import * as S from "./SearchBar.styles";
+import { Props } from "./SearchBar.types";
 
-interface IProps {
-  value: string;
-  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
-}
-
-const SearchBar = ({ value, onChange, ...rest }: IProps) => {
+const SearchBar = ({ value, onChange, ...rest }: Props) => {
   return (
     <>
       <S.SearchContainer>
@@ -29,42 +24,6 @@ const SearchBar = ({ value, onChange, ...rest }: IProps) => {
       </S.SearchContainer>
     </>
   );
-};
-
-const S = {
-  SearchContainer: styled.div`
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    width: 100%;
-    height: 100px;
-    background-color: ${(props) => props.theme.color.bgColor};
-  `,
-
-  SearchDiv: styled.div`
-    display: flex;
-    align-items: center;
-    min-width: 200px;
-    position: relative;
-    padding: 0 2rem;
-    @media screen and (max-width: ${media.mobile}) {
-      width: 100px;
-    }
-  `,
-
-  SearchIcon: styled.div`
-    display: flex;
-    position: absolute;
-    width: 1rem;
-    height: 1rem;
-    left: 9%;
-    background-size: cover;
-    background-image: url("/img/search.svg");
-
-    @media screen and (max-width: ${media.mobile}) {
-      left: 0;
-    }
-  `,
 };
 
 export default SearchBar;
