+++ src/components/SearchBar/SearchBar.types.ts
@@ -1,0 +1,4 @@
+export interface Props {
+  value: string;
+  onChange: (e: React.ChangeEvent<HTMLInputElement>) => void;
+}
