+++ src/components/SideBar/SideBar.tsx
@@ -1,32 +1,12 @@
-import { ReactNode } from "react";
-import { styled } from "styled-components";
-import { media } from "../../style/theme";
+import * as S from "./SideBar.styles";
+import { Props } from "./SideBar.types";
 
-interface IProps {
-  children: ReactNode;
-}
-
-function SideBar({ children }: IProps) {
+function SideBar({ children }: Props) {
   return (
     <>
       <S.Container>{children}</S.Container>;
     </>
   );
 }
-
-const S = {
-  Container: styled.div`
-    display: flex;
-    position: fixed;
-    align-items: center;
-    margin-top: 10px;
-    width: inherit;
-    height: 90vh;
-    border-right: 5px solid #d9d9d9;
-    border-top: 3px solid #d9d9d9;
-    border-color: ${(props) => props.theme.color.navBorder};
-    background-color: ${(props) => props.theme.color.bgColor};
-  `,
-};
 
 export default SideBar;
