+++ src/components/SideBar/SideBar.types.ts
@@ -1,0 +1,5 @@
+import { ReactNode } from "react";
+
+export interface Props {
+  children: ReactNode;
+}
