+++ src/pages/HomeModal/HomeModal.tsx
@@ -7,17 +7,26 @@ import {
   TemplateContainer,
 } from "../../style/styled_Home";
 import { useRecoilState } from "recoil";
-import { templateState } from "../../atom/atom";
-import ModalContent from "./ModalContent";
+import { albumState, templateState } from "../../atom/atom";
+import ModalContent, { IAlbum } from "./ModalContent";
 import ListContent from "./ListContent";
 import { C_FlexBox } from "../../style/styled_LogIn";
-import { AddButton } from "../../style/styled_Modal";
+import {
+  AddButton,
+  CustomButton,
+  UploadButton,
+} from "../../style/styled_Modal";
+import AlbumApi from "../../services/albumApi";
+import styled from "styled-components";
+import { FlexColumnCenterCSS } from "../../style/commonStyle";
+import { media } from "../../style/theme";
 
 function HomePage({
   setState,
 }: {
   setState: React.Dispatch<React.SetStateAction<boolean>>;
 }) {
+  const [board, setBoard] = useRecoilState<IAlbum>(albumState);
   const [modalState, setModalState] = useState<any>(null);
   const [listNum, setListNum] = useState(0);
   const [template, setTemplate] = useRecoilState(templateState);
@@ -25,9 +34,16 @@ function HomePage({
     setListNum(template.length);
   }, []);
 
+  const uploadClick = async () => {
+    const data = await AlbumApi.regist(
+      { pages: template },
+      { ...board, num: template.length }
+    );
+  };
+
   return (
-    <React.Fragment>
-      <Modal onClose={() => setState(false)}>
+    <Modal onClose={() => setState(false)}>
+      <S.Container>
         <ChoiceBox>
           <Phrase>템플릿</Phrase>
           <TemplateContainer>
@@ -78,9 +94,31 @@ function HomePage({
         <C_FlexBox>
           <ListContent />
         </C_FlexBox>
-      </Modal>
-    </React.Fragment>
+      </S.Container>
+      <S.ButtonWrapper>
+        <CustomButton onClick={uploadClick}>제출</CustomButton>
+      </S.ButtonWrapper>
+    </Modal>
   );
 }
+
+const S = {
+  Container: styled.div`
+    display: flex;
+    flex-direction: column;
+    position: relative;
+
+    @media screen and (max-width: ${media.mobile}) {
+      overflow-y: scroll;
+      width: 300px;
+      height: 500px;
+    }
+  `,
+  ButtonWrapper: styled.div`
+    position: absolute;
+    left: 35%;
+    z-index: 2;
+  `,
+};
 
 export default HomePage;
