+++ src/pages/HomeModal/ModalContent.tsx
@@ -6,6 +6,9 @@ import { useState } from "react";
 import { ButtonDiv, CustomButton } from "../../style/styled_Modal";
 import { ModalContainer, ModalImgDiv } from "../../style/styled_Home";
 import ModalInput from "./ModalInput";
+import { styled } from "styled-components";
+import { media } from "../../style/theme";
+import { FlexCenterCSS } from "../../style/commonStyle";
 
 export interface ITemplate {
   data: string;
@@ -112,21 +115,21 @@ function ModalContent({
               onChange={onFileReaderChange}
             />
           </CustomButton>
-          <ModalImgDiv>
+          <S.ImageWrapper>
             {fileReader ? (
               <img
-                style={{ maxWidth: "300px", maxHeight: "300px" }}
+                style={{ maxWidth: "100%", maxHeight: "100%" }}
                 src={fileReader}
                 alt="업로드 오류"
               />
             ) : (
               <img
-                style={{ width: "200px", height: "200px" }}
+                style={{ width: "100%", height: "100%" }}
                 src="/img/image.svg"
                 alt="이미지"
               />
             )}
-          </ModalImgDiv>
+          </S.ImageWrapper>
         </>
       ) : (
         <FileInput setText={setText}></FileInput>
@@ -148,5 +151,17 @@ function ModalContent({
     </ModalContainer>
   );
 }
+
+const S = {
+  ImageWrapper: styled.div`
+    width: 200px;
+    height: 200px;
+
+    @media screen and (max-width: ${media.mobile}) {
+      width: 30vw;
+      height: 30vw;
+    }
+  `,
+};
 
 export default ModalContent;
